+++ packages/bridge/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.2.0
+
+### Minor Changes
+
+- logger added
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors@0.3.0
+
 All notable changes to this project will be documented in this file.
 
 ## 0.1.0
@@ -7,4 +18,3 @@ All notable changes to this project will be documented in this file.
 ### Minor Changes
 
 - Initial package scaffold.
-