+++ packages/errors-nest/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.4.0
+
+### Minor Changes
+
+- logger added
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors@0.3.0
+
 ## 0.3.0
 
 ### Minor Changes